Get rid of kernel_user_comm.c
A similar functionnality is now handled inside liblustreapi_hsm.c.

+++ include/lustre/lustre.h
@@ -197,7 +197,7 @@ int llapi_hsm_copytool_register(const struct lustre_fs_h *lfsh,
 int llapi_hsm_copytool_unregister(struct hsm_copytool_private **priv);
 int llapi_hsm_copytool_get_fd(struct hsm_copytool_private *ct);
 int llapi_hsm_copytool_recv(struct hsm_copytool_private *priv,
-			    struct hsm_action_list **hal, int *msgsize);
+			    struct hsm_action_list **hal, size_t *msgsize);
 int llapi_hsm_action_begin(struct hsm_copyaction_private **phcp,
 			   const struct hsm_copytool_private *ct,
 			   const struct hsm_action_item *hai,
